Thermal control (#11)
* Properties for database

Loop that exists on exterior of station, cools through radiators to space. Radiators rotate in relation to the sun to provide optimal cooling

* Properties for database

Loop that exists inside the station. Interfaces with various cold plates to cool equipment and air. Circulates water on ISS.

* Initial controller/data loader concept 

Idea is emulate the firmware controller of the system by having a class emulate sensor readings, change hardware states (ie pump on/off) and then store that 'momentary' state into the database. The UI would then utilize this state data on the other side of the database.

* Commented on property/method intentions

* No longer using this class

* new commit on water-system branch

* newer commit on water system branch

* Changed namespace and directory

* changed namespace and directory

* changed namespace and directory

* changed directory for real

* misplaced duplicate

* no longer using this class

* Added property for overall component status

* Added property for overall system status

* added using directives, changed DI provider to OrbitServiceProvider

* Added internal and external cooling systems with value ranges

* Deleted classes no longer in use

* Fixed build problems

* Added tolerances, corrected wrong AlertLevel generation, adjusted methods